requirements.txt
pytz
numpy
PyYAML>=5.1.1
requests
websocket-client
locust
PyMySQL
pymongo
